Update util.cpp
For this new release 95610 we should change the default folder to .ioncoin (from .ionomy) so that as the community migrates to the new fork, they do not overwrite/modify their existing .ionomy data folder thus preserving their wallet.dat. It will also make the process of migrating to the new chain easier as community members will not need to delve in to their .ionomy data folder and begin to delete specific files and folders thus possibly and accidentally deleting their wallet.dat.

Former-commit-id: ec8d2b1fe03f5190d8ea1811167b0ceac194a858